Create the beginning of a Unit framework
Base units can be declared using the UnitSystem type by supplying a name
and an abbreviation. To start, the SI unit system is created with Meter
(m) and Second (s) as units. Right now, the type definition and some
behavior (printing, retrieving the power (m^2 for example)) for these
units is defined in a hard-coded loop. This loop should be generalized
into a macro.

Some ideas for how to handle composite units (e.g. m/s):
 - (Ab)use a type union (e.g. m/s = Union{Meter{1}, Second{-1}})
   Pro: built into language
   Con: will not type check, is an abstract type
 - Construct a type called CompositeUnit or similar
   Pro: easy
   Con: will not type check, must define that behavior and will be less
        efficient than language-enforced types, breaks pattern of each
        unit as its own type
 - Use Tuple types
   Pro: can type check at compiler stage
   Con: Not sure it will work