BIG CHANGE: units don't carry values
I decided units should be their own type independent of their values. I
think this has simplified things a lot. The functionality is the same as
in the previous commit, except the values are no longer tracked and
addition and subtraction are not defined anymore. I may define them
later.